+++ lib/index.d.ts
@@ -1,3 +1,4 @@
 export * from './jsxComponent';
 export * from './jsxPlugins';
 export * from './types';
+export * from './plugins';