+++ lib/plugins/context/testHandler.d.ts
@@ -1,0 +1,9 @@
+import { ContextProps } from './context';
+declare global {
+    namespace JSX {
+        interface IntrinsicElements {
+            context: ContextProps;
+        }
+    }
+}
+export declare const testHandler: import("innet").Handler;