+++ src/plugins/context/context.ts
@@ -1,0 +1,35 @@
+import innet, { Handler } from 'innet'
+
+import { useHandler } from '../../jsxComponent'
+import { JSXPluginElement } from '../../jsxPlugins'
+
+export interface ContextProps <D = any> {
+  for: Context<D>
+  set?: D
+}
+
+export function useContext <D = any, Def = D> (context: Context<D, Def>): D | Def {
+  return context.get(useHandler())
+}
+
+export class Context <D = any, Def = D> {
+  readonly key: string
+
+  constructor (public readonly defaultValue?: Def, name?: string) {
+    this.key = Symbol(name) as unknown as string
+  }
+
+  get (handler: Handler): D | Def {
+    return this.key in handler ? handler[this.key] : this.defaultValue
+  }
+}
+
+export function createContextHandler <D> (handler: Handler, context: Context<D>, value: D): Handler {
+  const childrenHandler = Object.create(handler)
+  childrenHandler[context.key] = value
+  return childrenHandler
+}
+
+export function context ({ props, children }: JSXPluginElement<ContextProps>, handler: Handler) {
+  return innet(children, createContextHandler(handler, props.for, props.set))
+}